refactor: split ai intent and smart pipelines
